Move dangerous Conversion to tests
